+++ src/main/java/com/kkh/board/repository/ReplyRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.kkh.board.model.Reply;
 
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+public interface ReplyRepository extends JpaRepository<Reply, Long> {
 
 	@Modifying
 	@Query(value="INSERT INTO reply(userId, boardId, content, createDate) VALUES(?1, ?2, ?3, now())", nativeQuery = true)
-	int mSave(int userId, int boardId, String content); // 업데이트된 행의 개수를 리턴해줌.
+	int mSave(Long userId, Long boardId, String content); // 업데이트된 행의 개수를 리턴해줌.
 }
